Properly log session IP address
When Openfire closes a session on version 4, the MINA IoSession
appears to 'go away' more quickly and the IP address associated
with the session is already lost by the time this plugin attempts
to log the change.

This change simply uses the already existing `userStatus` table
row for the history table.